Simulation d'une mission avec résultat à la fin. Affichage des différentes mission pour un apprenant. Possibilité de l'inscrire si on est admin. Code bizarre dans MissionService. La BD n'as pas du être générer via hibernate ou avec un autre naming. Compléter action scoreminimal et indicator valueIfCheck et unCheck si bug.

+++ src/main/java/com/epul/oeuvre/repositories/ApprenantRepository.java
@@ -3,5 +3,5 @@ package com.epul.oeuvre.repositories;
 import com.epul.oeuvre.domains.LearnerEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ApprenantRepository extends JpaRepository<LearnerEntity, Integer> {
+public interface ApprenantRepository extends JpaRepository<LearnerEntity, Long> {
 }
